Change structure.
